requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy

face_alignment
python_speech_features
numba
resampy
transformers
pyaudio
soundfile
einops
configargparse

lpips
imageio-ffmpeg
